refactor: 폴더 구조 정리 및 import 경로 수정
