+++ loneleap-client/src/components/Review/ReviewList.jsx
@@ -1,11 +1,11 @@
 // src/components/Review/ReviewList.jsx
 import { useState } from "react";
 
-import { useReviews } from "services/queries/useReviews";
+import { useReviews } from "services/queries/review/useReviews";
 
 import ReviewCard from "./ReviewCard";
-import LoadingSpinner from "components/LoadingSpinner";
-import EmptyState from "components/EmptyState";
+import LoadingSpinner from "components/common/LoadingSpinner";
+import EmptyState from "components/common/EmptyState";
 import ReviewSortButtons from "./ReviewSortButtons";
 
 export default function ReviewList() {
