+++ loneleap-client/src/pages/Auth/Login.jsx
@@ -1,5 +1,5 @@
 // src/pages/Auth/Login.jsx
-import AuthForm from "components/AuthForm";
+import AuthForm from "components/auth/AuthForm";
 
 export default function Login() {
   return (
