+++ loneleap-client/src/pages/Itinerary/Create.jsx
@@ -1,4 +1,4 @@
-import ItineraryForm from "components/ItineraryForm";
+import ItineraryForm from "components/itinerary/ItineraryForm";
 
 export default function CreateItineraryPage() {
   return (
