+++ loneleap-client/src/pages/Itinerary/Detail.jsx
@@ -1,4 +1,4 @@
-import ItineraryDetail from "components/ItineraryDetail";
+import ItineraryDetail from "components/itinerary/ItineraryDetail";
 
 export default function ItineraryDetailPage() {
   return <ItineraryDetail />;
