+++ loneleap-client/src/services/queries/chat/useCreateChatRoom.js
@@ -1,7 +1,7 @@
 // src/services/queries/useCreateChatRoom.js
 
 import { collection, addDoc, serverTimestamp } from "firebase/firestore";
-import { db } from "../firebase";
+import { db } from "../../firebase";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
 /**
