+++ loneleap-client/src/services/queries/chat/useMyChatRooms.js
@@ -8,7 +8,7 @@ import {
   orderBy,
   limit,
 } from "firebase/firestore";
-import { db } from "../firebase";
+import { db } from "../../firebase";
 
 export const useMyChatRooms = (uid, options = {}) => {
   return useQuery({
