+++ loneleap-client/src/services/queries/chat/useReportMessage.js
@@ -1,5 +1,5 @@
 import { addDoc, collection, serverTimestamp } from "firebase/firestore";
-import { db } from "../firebase";
+import { db } from "../../firebase";
 import { useSelector } from "react-redux";
 import { useMutation } from "@tanstack/react-query";
 import { getDocs, query, where } from "firebase/firestore";
