+++ loneleap-client/src/services/queries/itinerary/useItineraryDetail.js
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { fetchItineraryById } from "../firestore";
+import { fetchItineraryById } from "../../firestore";
 
 export const useItineraryDetail = (id) => {
   return useQuery({
