+++ loneleap-client/src/services/queries/itinerary/useMyItineraries.js
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
-import { fetchUserItineraries } from "../firestore";
-import { auth } from "../firebase";
+import { fetchUserItineraries } from "../../firestore";
+import { auth } from "../../firebase";
 import { useAuthState } from "react-firebase-hooks/auth";
 
 export const useMyItineraries = (options = {}) => {
