+++ loneleap-client/src/services/queries/review/useAddReview.js
@@ -1,6 +1,6 @@
 // src/services/queries/useAddReview.js
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { db, storage } from "../firebase";
+import { db, storage } from "../../firebase";
 import { addDoc, collection, serverTimestamp } from "firebase/firestore";
 import { getDownloadURL, ref, uploadBytes } from "firebase/storage";
 import { useNavigate } from "react-router-dom";
