+++ loneleap-client/src/services/queries/review/useReportReview.js
@@ -9,7 +9,7 @@ import {
   query,
 } from "firebase/firestore";
 import { getAuth } from "firebase/auth";
-import { db } from "../firebase";
+import { db } from "../../firebase";
 import { useSelector } from "react-redux";
 
 export const useReportReview = () => {
