+++ loneleap-client/src/services/queries/review/useReviewDetail.js
@@ -1,7 +1,7 @@
 // src/services/queries/useReviewDetail.js
 import { useQuery } from "@tanstack/react-query";
 import { doc, getDoc } from "firebase/firestore";
-import { db } from "../firebase";
+import { db } from "../../firebase";
 
 /**
  * 특정 리뷰의 상세 정보를 가져오는 커스텀 훅
