+++ loneleap-client/src/services/queries/review/useReviews.js
@@ -1,7 +1,7 @@
 // src/services/queries/useReviews.js
 import { useQuery } from "@tanstack/react-query";
 import { collection, getDocs, limit, orderBy, query } from "firebase/firestore";
-import { db } from "../firebase";
+import { db } from "../../firebase";
 
 export const useReviews = () => {
   return useQuery({
